--- pkg/models/cluster.go
@@ -1,7 +1,0 @@
-package models
-
-import ()
-
-type Cluster struct {
-	Name string
-}
