--- pkg/models/manifests_suite_test.go
@@ -1,13 +1,0 @@
-package models
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestInstaller(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "Manifests Suite")
-}
